requirements.txt
numpy==1.23.5
scikit-learn==1.2.2
opencv-python==4.9.0
scikit-image==0.22.0
plantcv==4.2.1
matplotlib==3.7.1
torch==2.2.0
pillow==9.0.1
segment-anything==1.0
tqdm==4.66.1